pridal som BorrowingsService, cakam na servisy Tomasa (books), aby som mohol dodat dalsie upravy
