+++ .idea/libraries/Maven__io_swagger_core_v3_swagger_core_2_1_12.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.swagger.core.v3:swagger-core:2.1.12">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/swagger/core/v3/swagger-core/2.1.12/swagger-core-2.1.12.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/swagger/core/v3/swagger-core/2.1.12/swagger-core-2.1.12-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/swagger/core/v3/swagger-core/2.1.12/swagger-core-2.1.12-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
