+++ .idea/libraries/Maven__org_springdoc_springdoc_openapi_common_1_6_6.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.springdoc:springdoc-openapi-common:1.6.6">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springdoc/springdoc-openapi-common/1.6.6/springdoc-openapi-common-1.6.6.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springdoc/springdoc-openapi-common/1.6.6/springdoc-openapi-common-1.6.6-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springdoc/springdoc-openapi-common/1.6.6/springdoc-openapi-common-1.6.6-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
