+++ .idea/libraries/Maven__org_webjars_swagger_ui_4_5_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.webjars:swagger-ui:4.5.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/swagger-ui/4.5.0/swagger-ui-4.5.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/swagger-ui/4.5.0/swagger-ui-4.5.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/swagger-ui/4.5.0/swagger-ui-4.5.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
