+++ .idea/libraries/Maven__org_webjars_webjars_locator_core_0_48.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.webjars:webjars-locator-core:0.48">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/webjars-locator-core/0.48/webjars-locator-core-0.48.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/webjars-locator-core/0.48/webjars-locator-core-0.48-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/webjars/webjars-locator-core/0.48/webjars-locator-core-0.48-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
